Macro assume::assume

macro_rules! assume {
    (unsafe: $cond:expr $(,)?) => { ... };
    (unsafe: $cond:expr, $fmt:expr $(, $($args:tt)*)?) => { ... };
    (unsafe: @unreachable $(,)?) => { ... };
    (unsafe: @unreachable, $fmt:expr $(, $($args:tt)*)?) => { ... };
    (unsafe: $($_:tt)*) => { ... };
    ($($_:tt)*) => { ... };
}

Assumes that the given condition is true.

This macro allows the expression of invariants in code. For example, one might assume! that an index is in bounds prior to indexing into a slice - this would allow the optimizer to remove the bounds checking entirely, under the promises of assume. In debug_assertion configurations the expression is checked. Otherwise, it is unchecked (but present).

Use @unreachable as the condition to assume the code path cannot be reached.

Because this expresses unchecked information, the act of assuming is inherently unsafe. The safe (i.e., runtime checked) alternative to this is the assert! macro. If the condition is @unreachable, the safe alternative to this is the unreachable! macro.

See the module level documentation for more.